refactor: specify core.utils
